extract backspace logic from math-wrapper
