+++ packages/math-input/src/components/input/math-wrapper.ts
@@ -7,18 +7,23 @@
 import $ from "jquery";
 
 import Key from "../../data/keys";
-import {MathFieldInterface} from "../../types";
+import {MathFieldInterface, MathFieldActionType} from "../../types";
 import keyTranslator from "../key-translator";
 
 import {CursorContext} from "./cursor-contexts";
+import handleBackspace from "./handle-backspace";
+import {
+    isFraction,
+    isParens,
+    isLeaf,
+    isNumerator,
+    isDenominator,
+    isSubScript,
+    isSuperScript,
+    maybeFindCommand,
+    maybeFindCommandBeforeParens,
+} from "./mathquill-helpers";
 import MQ from "./mathquill-instance";
-
-enum ActionType {
-    WRITE = "write",
-    CMD = "cmd",
-    KEYSTROKE = "keystroke",
-    MQ_END = 0,
-}
 
 const customKeyTranslator = {
     ...keyTranslator,
@@ -37,46 +42,6 @@ const NormalCommands = {
 
 const ArithmeticOperators = ["+", "-", "\\cdot", "\\times", "\\div"];
 const EqualityOperators = ["=", "\\neq", "<", "\\leq", ">", "\\geq"];
-
-const Numerals = ["0", "1", "2", "3", "4", "5", "6", "7", "8", "9"];
-const GreekLetters = ["\\theta", "\\pi"];
-const Letters = [
-    "A",
-    "B",
-    "C",
-    "D",
-    "E",
-    "F",
-    "G",
-    "H",
-    "I",
-    "J",
-    "K",
-    "L",
-    "M",
-    "N",
-    "O",
-    "P",
-    "Q",
-    "R",
-    "S",
-    "T",
-    "U",
-    "V",
-    "W",
-    "X",
-    "Y",
-    "Z",
-];
-
-// We only consider numerals, variables, and Greek Letters to be proper
-// leaf nodes.
-const ValidLeaves = [
-    ...Numerals,
-    ...GreekLetters,
-    ...Letters.map((letter) => letter.toLowerCase()),
-    ...Letters.map((letter) => letter.toUpperCase()),
-];
 
 const KeysForJumpContext = {
     [CursorContext.IN_PARENS]: "JUMP_OUT_PARENTHESES",
@@ -150,7 +115,7 @@ class MathWrapper {
         ) {
             this._handleJumpOut(cursor, key);
         } else if (key === "BACKSPACE") {
-            this._handleBackspace(cursor);
+            handleBackspace(this.mathField);
         } else if (key === "LEFT") {
             this._handleLeftArrow(cursor);
         } else if (key === "RIGHT") {
@@ -207,7 +172,7 @@ class MathWrapper {
                 // Unless that would leave us mid-command, in which case, we
                 // need to adjust and place the cursor inside the parens
                 // following the command.
-                const command = this._maybeFindCommand(cursor[MQ.L]);
+                const command = maybeFindCommand(cursor[MQ.L]);
                 if (command && command.endNode) {
                     // NOTE(charlie): endNode should definitely be \left(.
                     cursor.insLeftOf(command.endNode);
@@ -260,24 +225,6 @@ class MathWrapper {
     // All of the code below is super fragile.  Please be especially careful
     // when upgrading MathQuill.
 
-    _handleBackspaceInNthRoot(cursor) {
-        const isAtLeftEnd = cursor[MQ.L] === ActionType.MQ_END;
-
-        const isRootEmpty = this._isInsideEmptyNode(
-            cursor.parent.parent.blocks[0].ends,
-        );
-
-        if (isAtLeftEnd) {
-            this._selectNode(cursor.parent.parent, cursor);
-
-            if (isRootEmpty) {
-                this.mathField.keystroke("Backspace");
-            }
-        } else {
-            this.mathField.keystroke("Backspace");
-        }
-    }
-
     /**
      * Advances the cursor to the next logical position.
      *
@@ -308,8 +255,8 @@ class MathWrapper {
                 // Find the nearest fraction to the right of the cursor.
                 let fractionNode;
                 let visitor = cursor;
-                while (visitor[MQ.R] !== ActionType.MQ_END) {
-                    if (this._isFraction(visitor[MQ.R])) {
+                while (visitor[MQ.R] !== MathFieldActionType.MQ_END) {
+                    if (isFraction(visitor[MQ.R])) {
                         fractionNode = visitor[MQ.R];
                     }
                     visitor = visitor[MQ.R];
@@ -344,7 +291,7 @@ class MathWrapper {
                 // Navigate right once more, if we're right before parens. This
                 // is to handle the standard case in which the subscript is the
                 // base of a custom log.
-                if (this._isParens(cursor[MQ.R])) {
+                if (isParens(cursor[MQ.R])) {
                     this.mathField.keystroke("Right");
                 }
                 break;
@@ -362,51 +309,6 @@ class MathWrapper {
         }
     }
 
-    /**
-     * Selects and deletes part of the expression based on the cursor location.
-     * See inline comments for precise behavior of different cases.
-     *
-     * @param {cursor} cursor
-     * @private
-     */
-    _handleBackspace(cursor) {
-        if (!cursor.selection) {
-            const parent = cursor.parent;
-            const grandparent = parent.parent;
-            const leftNode = cursor[MQ.L];
-
-            if (this._isFraction(leftNode)) {
-                this._selectNode(leftNode, cursor);
-            } else if (this._isSquareRoot(leftNode)) {
-                this._selectNode(leftNode, cursor);
-            } else if (this._isNthRoot(leftNode)) {
-                this._selectNode(leftNode, cursor);
-            } else if (this._isNthRootIndex(parent)) {
-                this._handleBackspaceInRootIndex(cursor);
-            } else if (leftNode.ctrlSeq === "\\left(") {
-                this._handleBackspaceOutsideParens(cursor);
-            } else if (grandparent.ctrlSeq === "\\left(") {
-                this._handleBackspaceInsideParens(cursor);
-            } else if (this._isInsideLogIndex(cursor)) {
-                this._handleBackspaceInLogIndex(cursor);
-            } else if (
-                leftNode.ctrlSeq === "\\ge " ||
-                leftNode.ctrlSeq === "\\le "
-            ) {
-                this._handleBackspaceAfterLigaturedSymbol(cursor);
-            } else if (
-                this._isNthRoot(grandparent) &&
-                leftNode === ActionType.MQ_END
-            ) {
-                this._handleBackspaceInNthRoot(cursor);
-            } else {
-                this.mathField.keystroke("Backspace");
-            }
-        } else {
-            this.mathField.keystroke("Backspace");
-        }
-    }
-
     _handleLeftArrow(cursor) {
         // If we're inside a function, and just after the left parentheses, we
         // need to skip the entire function name, rather than move the cursor
@@ -418,11 +320,11 @@ class MathWrapper {
         // the ActionType.MQ_END node, that our grandparent is the left parenthesis, and
         // the nodes to the left of our grandparent comprise a valid function
         // name.
-        if (cursor[MQ.L] === ActionType.MQ_END) {
+        if (cursor[MQ.L] === MathFieldActionType.MQ_END) {
             const parent = cursor.parent;
             const grandparent = parent.parent;
             if (grandparent.ctrlSeq === "\\left(") {
-                const command = this._maybeFindCommandBeforeParens(grandparent);
+                const command = maybeFindCommandBeforeParens(grandparent);
                 if (command) {
                     cursor.insLeftOf(command.startNode);
                     return;
@@ -435,7 +337,7 @@ class MathWrapper {
     }
 
     _handleRightArrow(cursor) {
-        const command = this._maybeFindCommand(cursor[MQ.R]);
+        const command = maybeFindCommand(cursor[MQ.R]);
         if (command) {
             // Similarly, if a function is to our right, then we need to place
             // the cursor at the start of its parenthetical content, which is
@@ -457,7 +359,7 @@ class MathWrapper {
 
         const precedingNode = cursor[MQ.L];
         const shouldPrefixWithParens =
-            precedingNode === ActionType.MQ_END ||
+            precedingNode === MathFieldActionType.MQ_END ||
             invalidPrefixes.includes(precedingNode.ctrlSeq.trim());
         if (shouldPrefixWithParens) {
             this.mathField.write("\\left(\\right)");
@@ -490,390 +392,13 @@ class MathWrapper {
         }
     }
 
-    /**
-     * Return the start node, end node, and full name of the command of which
-     * the initial node is a part, or `null` if the node is not part of a
-     * command.
-     *
-     * @param {node} initialNode - the node to included as part of the command
-     * @returns {null|object} - `null` or an object containing the start node
-     *                          (`startNode`), end node (`endNode`), and full
-     *                          name (`name`) of the command
-     * @private
-     */
-    _maybeFindCommand(initialNode) {
-        if (!initialNode) {
-            return null;
-        }
-
-        // MathQuill stores commands as separate characters so that
-        // users can delete commands one character at a time.  We iterate over
-        // the nodes from right to left until we hit a sequence starting with a
-        // '\\', which signifies the start of a command; then we iterate from
-        // left to right until we hit a '\\left(', which signifies the end of a
-        // command.  If we encounter any character that doesn't belong in a
-        // command, we return null.  We match a single character at a time.
-        // Ex) ['\\l', 'o', 'g ', '\\left(', ...]
-        const commandCharRegex = /^[a-z]$/;
-        const commandStartRegex = /^\\[a-z]$/;
-        const commandEndSeq = "\\left(";
-
-        // Note: We allowlist the set of valid commands, since relying solely on
-        // a command being prefixed with a backslash leads to undesired
-        // behavior. For example, Greek symbols, left parentheses, and square
-        // roots all get treated as commands.
-        const validCommands = ["\\log", "\\ln", "\\cos", "\\sin", "\\tan"];
-
-        let name = "";
-        let startNode;
-        let endNode;
-
-        // Collect the portion of the command from the current node, leftwards
-        // until the start of the command.
-        let node = initialNode;
-        while (node !== 0) {
-            const ctrlSeq = node.ctrlSeq.trim();
-            if (commandCharRegex.test(ctrlSeq)) {
-                name = ctrlSeq + name;
-            } else if (commandStartRegex.test(ctrlSeq)) {
-                name = ctrlSeq + name;
-                startNode = node;
-                break;
-            } else {
-                break;
-            }
-
-            node = node[MQ.L];
-        }
-
-        // If we hit the start of a command, then grab the rest of it by
-        // iterating rightwards to compute the full name of the command, along
-        // with its terminal node.
-        if (startNode) {
-            // Next, iterate from the start to the right.
-            node = initialNode[MQ.R];
-            while (node !== 0) {
-                const ctrlSeq = node.ctrlSeq.trim();
-                if (commandCharRegex.test(ctrlSeq)) {
-                    // If we have a single character, add it to the command
-                    // name.
-                    name = name + ctrlSeq;
-                } else if (ctrlSeq === commandEndSeq) {
-                    // If we hit the command end delimiter (the left
-                    // parentheses surrounding its arguments), stop.
-                    endNode = node;
-                    break;
-                }
-
-                node = node[MQ.R];
-            }
-            if (validCommands.includes(name)) {
-                return {name, startNode, endNode};
-            } else {
-                return null;
-            }
-        } else {
-            return null;
-        }
-    }
-
-    /**
-     * Return the start node, end node, and full name of the command to the left
-     * of `\\left(`, or `null` if there is no command.
-     *
-     * @param {node} leftParenNode - node where .ctrlSeq == `\\left(`
-     * @returns {null|object} - `null` or an object containing the start node
-     *                          (`startNode`), end node (`endNode`), and full
-     *                          name (`name`) of the command
-     * @private
-     */
-    _maybeFindCommandBeforeParens(leftParenNode) {
-        return this._maybeFindCommand(leftParenNode[MQ.L]);
-    }
-
-    _selectNode(node, cursor) {
-        cursor.insLeftOf(node);
-        cursor.startSelection();
-        cursor.insRightOf(node);
-        cursor.select();
-        cursor.endSelection();
-    }
-
-    _isFraction(node) {
-        return node.jQ && node.jQ.hasClass("mq-fraction");
-    }
-
-    _isNumerator(node) {
-        return node.jQ && node.jQ.hasClass("mq-numerator");
-    }
-
-    _isDenominator(node) {
-        return node.jQ && node.jQ.hasClass("mq-denominator");
-    }
-
-    _isSubScript(node) {
-        // NOTE(charlie): MyScript has a structure whereby its superscripts seem
-        // to be represented as a parent node with 'mq-sup-only' containing a
-        // single child with 'mq-sup'.
-        return (
-            node.jQ &&
-            (node.jQ.hasClass("mq-sub-only") || node.jQ.hasClass("mq-sub"))
-        );
-    }
-
-    _isSuperScript(node) {
-        // NOTE(charlie): MyScript has a structure whereby its superscripts seem
-        // to be represented as a parent node with 'mq-sup-only' containing a
-        // single child with 'mq-sup'.
-        return (
-            node.jQ &&
-            (node.jQ.hasClass("mq-sup-only") || node.jQ.hasClass("mq-sup"))
-        );
-    }
-
-    _isParens(node) {
-        return node && node.ctrlSeq === "\\left(";
-    }
-
-    _isLeaf(node) {
-        return (
-            node && node.ctrlSeq && ValidLeaves.includes(node.ctrlSeq.trim())
-        );
-    }
-
-    _isSquareRoot(node) {
-        return (
-            node.blocks &&
-            node.blocks[0].jQ &&
-            node.blocks[0].jQ.hasClass("mq-sqrt-stem")
-        );
-    }
-
-    _isNthRoot(node) {
-        return (
-            node.blocks &&
-            node.blocks[0].jQ &&
-            node.blocks[0].jQ.hasClass("mq-nthroot")
-        );
-    }
-
-    _isNthRootIndex(node) {
-        return node.jQ && node.jQ.hasClass("mq-nthroot");
-    }
-
-    _isInsideLogIndex(cursor) {
-        const grandparent = cursor.parent.parent;
-
-        if (grandparent && grandparent.jQ.hasClass("mq-supsub")) {
-            const command = this._maybeFindCommandBeforeParens(grandparent);
-
-            if (command && command.name === "\\log") {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    _isInsideEmptyNode(cursor) {
-        return (
-            cursor[MQ.L] === ActionType.MQ_END &&
-            cursor[MQ.R] === ActionType.MQ_END
-        );
-    }
-
-    _handleBackspaceInRootIndex(cursor) {
-        if (this._isInsideEmptyNode(cursor)) {
-            // When deleting the index in a nthroot, we change from the nthroot
-            // to a sqrt, e.g. \sqrt[|]{35x-5} => |\sqrt{35x-5}.  If there's no
-            // content under the root, then we delete the whole thing.
-
-            const grandparent = cursor.parent.parent;
-            const latex = grandparent.latex();
-            const reinsertionPoint = grandparent[MQ.L];
-
-            this._selectNode(grandparent, cursor);
-
-            const rootIsEmpty = grandparent.blocks[1].jQ.text() === "";
-
-            if (rootIsEmpty) {
-                // If there is not content under the root then simply delete
-                // the whole thing.
-                this.mathField.keystroke("Backspace");
-            } else {
-                // Replace the nthroot with a sqrt if there was content under
-                // the root.
-
-                // Start by deleting the selection.
-                this.mathField.keystroke("Backspace");
-
-                // Replace the nth-root with a sqrt.
-                this.mathField.write(latex.replace(/^\\sqrt\[\]/, "\\sqrt"));
-
-                // Adjust the cursor to be to the left the sqrt.
-                if (reinsertionPoint === ActionType.MQ_END) {
-                    this.mathField.moveToDirEnd(MQ.L);
-                } else {
-                    cursor.insRightOf(reinsertionPoint);
-                }
-            }
-        } else {
-            if (cursor[MQ.L] !== ActionType.MQ_END) {
-                // If the cursor is not at the leftmost position inside the
-                // root's index, delete a character.
-                this.mathField.keystroke("Backspace");
-            } else {
-                // TODO(kevinb) verify that we want this behavior after testing
-                // Do nothing because we haven't completely deleted the
-                // index of the radical.
-            }
-        }
-    }
-
-    _handleBackspaceInLogIndex(cursor) {
-        if (this._isInsideEmptyNode(cursor)) {
-            const grandparent = cursor.parent.parent;
-            const command = this._maybeFindCommandBeforeParens(grandparent);
-
-            cursor.insLeftOf(command?.startNode);
-            cursor.startSelection();
-
-            if (grandparent[MQ.R] !== ActionType.MQ_END) {
-                cursor.insRightOf(grandparent[MQ.R]);
-            } else {
-                cursor.insRightOf(grandparent);
-            }
-
-            cursor.select();
-            cursor.endSelection();
-
-            const isLogBodyEmpty = grandparent[MQ.R].contentjQ.text() === "";
-
-            if (isLogBodyEmpty) {
-                // If there's no content inside the log's parens then delete the
-                // whole thing.
-                this.mathField.keystroke("Backspace");
-            }
-        } else {
-            this.mathField.keystroke("Backspace");
-        }
-    }
-
-    _handleBackspaceOutsideParens(cursor) {
-        // In this case the node with '\\left(' for its ctrlSeq
-        // is the parent of the expression contained within the
-        // parentheses.
-        //
-        // Handle selecting an expression before deleting:
-        // (x+1)| => |(x+1)|
-        // \log(x+1)| => |\log(x+1)|
-
-        const leftNode = cursor[MQ.L];
-        const rightNode = cursor[MQ.R];
-        const command = this._maybeFindCommandBeforeParens(leftNode);
-
-        if (command && command.startNode) {
-            // There's a command before the parens so we select it as well as
-            // the parens.
-            cursor.insLeftOf(command.startNode);
-            cursor.startSelection();
-            if (rightNode === ActionType.MQ_END) {
-                cursor.insAtRightEnd(cursor.parent);
-            } else {
-                cursor.insLeftOf(rightNode);
-            }
-            cursor.select();
-            cursor.endSelection();
-        } else {
-            cursor.startSelection();
-            cursor.insLeftOf(leftNode); // left of \\left(
-            cursor.select();
-            cursor.endSelection();
-        }
-    }
-
-    _handleBackspaceInsideParens(cursor) {
-        // Handle situations when the cursor is inside parens or a
-        // command that uses parens, e.g. \log() or \tan()
-        //
-        // MathQuill represents log(x+1) in roughly the following way
-        // [l, o, g, \\left[parent:[x, +, 1]]]
-        //
-        // If the cursor is inside the parentheses it's next to one of:
-        // x, +, or 1.  This makes sub_sub_expr its parent and sub_expr
-        // it's parent.
-        //
-        // Interestingly parent doesn't have any nodes to the left or
-        // right of it (even though the corresponding DOM node has
-        // ( and ) characters on either side.
-        //
-        // The grandparent's ctrlSeq is `\\left(`. The `\\right)` isn't
-        // stored anywhere.  NOTE(kevinb): I believe this is because
-        // MathQuill knows what the close paren should be and does the
-        // right thing at render time.
-        //
-        // This conditional branch handles the following cases:
-        // - \log(x+1|) => \log(x+|)
-        // - \log(|x+1) => |\log(x+1)|
-        // - \log(|) => |
-
-        if (cursor[MQ.L] !== ActionType.MQ_END) {
-            // This command contains math and there's some math to
-            // the left of the cursor that we should delete normally
-            // before doing anything special.
-            this.mathField.keystroke("Backspace");
-            return;
-        }
-
-        const grandparent = cursor.parent.parent;
-
-        // If the cursors is inside the parens at the start but the command
-        // has a subscript as is the case in log_n then move the cursor into
-        // the subscript, e.g. \log_{5}(|x+1) => \log_{5|}(x+1)
-
-        if (grandparent[MQ.L].sub) {
-            // if there is a subscript
-            if (grandparent[MQ.L].sub.jQ.text()) {
-                // and it contains text
-                // move the cursor to the right end of the subscript
-                cursor.insAtRightEnd(grandparent[MQ.L].sub);
-                return;
-            }
-        }
-
-        // Determine if the parens are empty before we modify the
-        // cursor's position.
-        const isEmpty = this._isInsideEmptyNode(cursor);
-
-        // Insert the cursor to the left of the command if there is one
-        // or before the '\\left(` if there isn't
-        const command = this._maybeFindCommandBeforeParens(grandparent);
-
-        cursor.insLeftOf((command && command.startNode) || grandparent);
-        cursor.startSelection();
-        cursor.insRightOf(grandparent);
-        cursor.select();
-        cursor.endSelection();
-
-        // Delete the selection, but only if the parens were empty to
-        // begin with.
-        if (isEmpty) {
-            this.mathField.keystroke("Backspace");
-        }
-    }
-
-    _handleBackspaceAfterLigaturedSymbol(cursor) {
-        this.mathField.keystroke("Backspace");
-        this.mathField.keystroke("Backspace");
-    }
-
     contextForCursor(cursor) {
         // First, try to find any fraction to the right, unimpeded.
         let visitor = cursor;
-        while (visitor[MQ.R] !== ActionType.MQ_END) {
-            if (this._isFraction(visitor[MQ.R])) {
+        while (visitor[MQ.R] !== MathFieldActionType.MQ_END) {
+            if (isFraction(visitor[MQ.R])) {
                 return CursorContext.BEFORE_FRACTION;
-            } else if (!this._isLeaf(visitor[MQ.R])) {
+            } else if (!isLeaf(visitor[MQ.R])) {
                 break;
             }
             visitor = visitor[MQ.R];
@@ -881,15 +406,15 @@ class MathWrapper {
 
         // If that didn't work, check if the parent or grandparent is a special
         // context, so that we can jump outwards.
-        if (this._isParens(cursor.parent && cursor.parent.parent)) {
+        if (isParens(cursor.parent && cursor.parent.parent)) {
             return CursorContext.IN_PARENS;
-        } else if (this._isNumerator(cursor.parent)) {
+        } else if (isNumerator(cursor.parent)) {
             return CursorContext.IN_NUMERATOR;
-        } else if (this._isDenominator(cursor.parent)) {
+        } else if (isDenominator(cursor.parent)) {
             return CursorContext.IN_DENOMINATOR;
-        } else if (this._isSubScript(cursor.parent)) {
+        } else if (isSubScript(cursor.parent)) {
             return CursorContext.IN_SUB_SCRIPT;
-        } else if (this._isSuperScript(cursor.parent)) {
+        } else if (isSuperScript(cursor.parent)) {
             return CursorContext.IN_SUPER_SCRIPT;
         } else {
             return CursorContext.NONE;
